Budget Detail screen expense grouping utility function create and implement
