+++ apps/mobile/app/screens/RecurringExpenseDetailScreen.tsx
@@ -14,7 +14,7 @@ import {
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { formatDate } from '../utils/dateUtils';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import EditExpenseModal from '../components/expense/EditExpenseModal';
+import EditExpenseModal from '../components/expense/expenseForms/EditExpenseModal';
 import { RecurringExpense } from '../types/recurringExpense';
 import {
   deleteRecurringExpense,
